backend/requirements.txt
chromadb==1.0.4
chroma-hnswlib==0.7.6
python-dotenv==1.0.0
google-genai